+++ tests/test_utils/test_threading_utils.py
@@ -2,7 +2,6 @@
 """
 Tests for threading related utilities.
 """
-import pytest
 from unittest.mock import patch
 
 from nessai.utils.threading import configure_threads
@@ -23,10 +22,3 @@ def test_configure_threads_none():
     with patch("torch.set_num_threads") as mock:
         configure_threads(pytorch_threads=None)
     mock.assert_not_called()
-
-
-def test_max_threads_warning():
-    """Assert a deprecation warning is raised if max threads is specified"""
-    with pytest.warns(DeprecationWarning) as record:
-        configure_threads(max_threads=1)
-    assert "`max_threads` is deprecated" in str(record[0].message)